+++ js/components/_index.js
@@ -1,7 +1,0 @@
-export * from './projectNameInput.js';
-export * from './frameworkSelector.js';
-export * from './dependenciesSelector.js';
-export * from './licenseSelector.js';
-export * from './optionalToggle.js';
-export * from './generateButton.js';
-export * from './previewSection.js';