+++ js/components/generateButton.js
@@ -1,6 +1,0 @@
-
-export const generateButton = () => `
-  <div class="form-group">
-    <button id="generate-btn">Generate Project</button>
-  </div>
-`;